CRUD_listing

+++ client/src/pages/Profile.jsx
@@ -30,6 +30,8 @@ export default function Profile() {
     const [fileUploadError, setFileUploadError] = useState(false);
     const [formData, setFormData] = useState({});
     const [updateSuccess, setUpdateSuccess] = useState(false);
+    const [showListingsError, setShowListingsError] = useState(false);
+    const [userListings, setUserListings] = useState([]);
     const dispatch = useDispatch();
 
     //firebase storage
@@ -97,6 +99,9 @@ export default function Profile() {
     };
 
     const handleDeleteUser = async () => {
+        if (!window.confirm('Are you sure you want to delete your account?'))
+            return;
+
         try {
             dispatch(deleteUserStart());
             const res = await fetch(`/api/user/delete/${currentUser._id}`, {
@@ -129,6 +134,73 @@ export default function Profile() {
             dispatch(logOutUserSuccess(data));
         } catch (error) {
             dispatch(logOutUserFailure(error.message));
+        }
+    };
+
+    const handleShowListings = async () => {
+        try {
+            setShowListingsError(false);
+            const res = await fetch(`/api/user/listings/${currentUser._id}`);
+            const data = await res.json();
+            if (data.success === false) {
+                setShowListingsError(true);
+                return;
+            }
+
+            setUserListings(data);
+        } catch (error) {
+            setShowListingsError(true);
+        }
+    };
+
+    const handleListingDelete = async (listingId) => {
+        if (
+            !window.confirm(
+                `Are you sure you want to delete your listing? ${listingId.name}`,
+            )
+        )
+            return;
+
+        try {
+            const res = await fetch(`/api/listing/delete/${listingId._id}`, {
+                method: 'DELETE',
+            });
+            const data = await res.json();
+            if (data.success === false) {
+                console.log(data.message);
+                return;
+            }
+
+            setUserListings((prev) =>
+                prev.filter((listing) => listing._id !== listingId),
+            );
+        } catch (error) {
+            console.log(error.message);
+        }
+    };
+
+    const handleListingUpdate = async (listingId) => {
+        try {
+            const res = await fetch(`/api/listing/update/${listingId}`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify(formData),
+            });
+
+            const data = await res.json();
+
+            if (data.success === false) {
+                console.log(data.message);
+                return;
+            }
+
+            setUserListings((prev) =>
+                prev.filter((listing) => listing._id !== listingId),
+            );
+        } catch (error) {
+            console.log(error.message);
         }
     };
 
@@ -223,6 +295,59 @@ export default function Profile() {
             <p className="text-green-700 mt-5">
                 {updateSuccess ? 'User is updated successfully' : ''}{' '}
             </p>
+            <button
+                onClick={handleShowListings}
+                className="text-green-700 w-full"
+            >
+                Show listing
+            </button>
+
+            {/* error listing */}
+            <p className="text-red-700 mt-5">
+                {showListingsError ? 'Error showing listings' : ''}
+            </p>
+
+            {userListings && userListings.length > 0 && (
+                <div className="flex flex-col gap-4">
+                    <h1 className="text-center mt-7 text-2xl font-semibold">
+                        Your Listings
+                    </h1>
+                    {userListings.map((listing) => (
+                        <div
+                            key={listing._id}
+                            className="border rounded-lg p-3 flex justify-between items-center gap-4"
+                        >
+                            <Link to={`/listing/${listing._id}`}>
+                                <img
+                                    src={listing.imageUrls[0]}
+                                    alt="listing cover"
+                                    className="h-16 w-16 object-contain"
+                                />
+                            </Link>
+                            <Link
+                                className="text-slate-700 font-semibold  hover:underline truncate flex-1"
+                                to={`/listing/${listing._id}`}
+                            >
+                                <p>{listing.name}</p>
+                            </Link>
+
+                            <div className="flex flex-col item-center">
+                                <button
+                                    onClick={() => handleListingDelete(listing)}
+                                    className="text-red-700 uppercase"
+                                >
+                                    Delete
+                                </button>
+                                <Link to={`/update-listing/${listing._id}`}>
+                                    <button className="text-green-700 uppercase">
+                                        Edit
+                                    </button>
+                                </Link>
+                            </div>
+                        </div>
+                    ))}
+                </div>
+            )}
         </div>
     );
 }
